ItemInfo 추가 및 필요 cs파일과 Data 설정 진행 중

+++ Assets/Script/GameData/Model/ItemModel.cs
@@ -7,7 +7,7 @@ namespace Script.GameData.Model {
         [Key(0)] public long uid;
 
         //아이템 정보의 uid
-        [Key(1)] public long   infoUid;
+        [Key(1)] public int    infoUid;
         [Key(2)] public double count;
 
         [Key(3)] public int level;
